requirements.txt
pandas
streamlit
plotly
google-generativeai
reportlab
python-dotenv
scikit-learn
langdetect
requests